requirements.txt
amqp==5.2.0
asgiref==3.8.1
asttokens==3.0.0
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
bcrypt==4.2.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.1
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.1
daphne==4.1.2
decorator==5.2.1
Django==4.2
django-extensions==3.2.3
docker==7.1.0
executing==2.1.0
graphviz==0.20.3
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
ipaddr==2.2.0
ipython==8.31.0
jedi==0.19.2
Jinja2==3.1.5
jsonpickle==4.0.1
kombu==5.4.2
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
msgpack==1.1.0
networkx==3.4.2
paramiko==3.5.0
parso==0.8.4
pillow==10.4.0
prompt_toolkit==3.0.48
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
PyMySQL==1.1.1
PyNaCl==1.5.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyvis==0.3.2
pywin32==308
PyYAML==6.0.2
redis==5.1.1
requests==2.32.3
service-identity==24.1.0
setuptools==75.3.0
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
tftpy==0.8.2
traitlets==5.14.3
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.1.0
